+++ promotion/user-service/src/main/java/com/study/userservice/repository/UserLoginHistoryRepository.java
@@ -5,6 +5,6 @@ import com.study.userservice.entity.UserLoginHistory;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserLoginHistoryRepository extends JpaRepository<UserLoginHistory, Integer> {
+public interface UserLoginHistoryRepository extends JpaRepository<UserLoginHistory, Long> {
   List<UserLoginHistory> findByUserOrderByLoginTimeDesc(User user);
 }
